requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
catboost
shap
jupyter
nbconvert
pyyaml
